Every RELACSPlugin got its own EventList whose indices are updated whenever the plugin needs new data.
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@1045 1550813e-0d35-0410-aec7-fc86b090456a

+++ plugins/auditory/src/adaptedficurves.cc
@@ -406,7 +406,7 @@ void AdaptedFICurves::analyze( EventList &spikes, SampleDataD &rate,
 	       signalTime() );
 
   // spike count:
-  spikecount.push( count, spikes.back().count( 0.0, duration ) );
+  spikecount.push( count, double( spikes.back().count( 0.0, duration ) ) );
 
   // firing frequency:
   SampleDataD ratesd( rate );
